backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
websockets==12.0
email-validator==2.1.0.post1
cryptography==42.0.0
beautifulsoup4==4.12.3
requests==2.31.0
dnspython==2.4.2
selenium==4.16.0
webdriver-manager==4.0.1
openpyxl==3.1.2

# Authentication dependencies
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

